-revise solar_system.yaml

+++ src/datastore.py
@@ -176,8 +176,8 @@ def _latitude(rows=4, cols=8, radius=1, offset=False):
 
 
 def _oblate_sphere(rows=4, cols=None, radius=(1200 * u.km,) * 3, offset=False):
-    verts = np.empty((rows + 1, cols + 1, 3), dtype=np.float32)
-    tcrds = np.empty((rows + 1, cols + 1, 2), dtype=np.float32)
+    verts = np.empty((rows + 1, cols + 1, 3), dtype=np.float64)
+    tcrds = np.empty((rows + 1, cols + 1, 2), dtype=np.float64)
     norms = np.linalg.norm(verts)
 
     # compute vertices
